+++ src/geomath.h
@@ -45,11 +45,16 @@ int bearing_position(const double lat, const double lon,
 double haversine(const double lat1, const double lon1,
                  const double lat2, const double lon2);
 double karney_distance(double lat1, double lon1, double lat2, double lon2);
+double karney_bearing(const double lat1, const double lon1,
+                      const double lat2, const double lon2);
 double distance(const DistFormula formula,
                 const double lat1, const double lon1,
                 const double lat2, const double lon2);
 double initial_bearing(const double lat1, const double lon1,
                        const double lat2, const double lon2);
+double bearing(const DistFormula formula,
+               const double lat1, const double lon1,
+               const double lat2, const double lon2);
 void rand_pos(double *dlat, double *dlon,
               const double c_lat, const double c_lon,
               const double maxdist, const double mindist);